@@ -39,7 +39,7 @@ export class GroupComponent implements OnInit, OnDestroy {
   batchInfos = new Map();
   isLdapMode: boolean;
 
-  @ViewChild(AddGroupModalComponent) newGroupModal: AddGroupModalComponent;
+  @ViewChild(AddGroupModalComponent, {static: false}) newGroupModal: AddGroupModalComponent;
 
   constructor(
     private operationService: OperationService,
