@@ -85,11 +85,7 @@ export function getCurrentLanguage(translateService: TranslateService) {
         deps: [ AppConfigService, SkinableConfig],
         multi: true
       },
-      {
-        provide: LOCALE_ID,
-        useFactory: getCurrentLanguage,
-        deps: [ TranslateService ]
-      }
+      {provide: LOCALE_ID, useValue: "en-US"}
     ],
     bootstrap: [AppComponent]
 })
